+++ src/esrf_ontologies/technique/__init__.py
@@ -3,7 +3,6 @@ from typing import Generator
 from typing import List
 from typing import Optional
 from typing import Set
-from typing import Tuple
 
 from ..db import load_techniques as _load_techniques
 from .types import BLISS_SCANINFO_CATEGORY  # noqa F401
@@ -11,20 +10,20 @@ from .types import Technique
 from .types import TechniqueMetadata
 
 
-def get_technique_metadata(*names: Tuple[str]) -> TechniqueMetadata:
+def get_technique_metadata(*identifiers: str) -> TechniqueMetadata:
     """Returns an object that can generate several types of metadata
-    associated to the provided technique names."""
-    return TechniqueMetadata(techniques=set(_iter_from_names(*names)))
+    associated to the provided technique names or IRIs."""
+    return TechniqueMetadata(techniques=set(_iter_from_identifiers(*identifiers)))
 
 
-def get_techniques(*names: Tuple[str]) -> Set[Technique]:
-    """Returns a set of techniques referenced by the provided technique names."""
-    return set(_iter_from_names(*names))
+def get_techniques(*identifiers: str) -> Set[Technique]:
+    """Returns a set of techniques referenced by the provided technique names or IRIs."""
+    return set(_iter_from_identifiers(*identifiers))
 
 
-def get_technique(name: str) -> Technique:
-    """Returns a technique referenced by the provided technique name."""
-    return next(_iter_from_names(name))
+def get_technique(identifier: str) -> Technique:
+    """Returns a technique referenced by the provided technique name or IRI."""
+    return next(_iter_from_identifiers(identifier))
 
 
 def get_ontology_version(metadata: Optional[TechniqueMetadata] = None) -> str:
@@ -48,19 +47,25 @@ def get_all_techniques() -> List[Technique]:
             names=tuple(technique["names"]),
             description=technique["description"],
             ontology_version=technique["ontology_version"],
+            versioned_iri=technique["versioned_iri"],
         )
         for ontology_name in ["ESRFET"]
         for technique in _load_techniques(ontology_name)
     ]
 
 
-def _iter_from_names(*names: Tuple[str]) -> Generator[Technique, None, None]:
+def _iter_from_identifiers(*identifiers: str) -> Generator[Technique, None, None]:
     techniques = get_all_techniques()
-    for name in sorted(set(names)):
+    for identifier in sorted(set(identifiers)):
+        identifier_lower = identifier.lower()
         for technique in techniques:
             technique_names = set(map(str.lower, technique.names))
-            if name.lower() in technique_names:
+            if (
+                identifier_lower in technique_names
+                or identifier == technique.iri
+                or identifier == technique.versioned_iri
+            ):
                 yield technique
                 break
         else:
-            raise KeyError(f"'{name}' is not a known technique name.")
+            raise KeyError(f"'{identifier}' is not a known technique name or IRI.")