+++ src/esrf_ontologies/tests/test_techniques.py
@@ -24,6 +24,19 @@ def test_get_technique():
     assert technique in all_techniques
 
 
+def test_get_technique_by_iri():
+    all_techniques = set(get_all_techniques())
+    technique = get_technique("https://w3id.org/PaN/ESRFET#XRF")
+    assert technique in all_techniques
+
+
+def test_get_technique_by_versioned_iri():
+    all_techniques = set(get_all_techniques())
+    version = get_ontology_version()
+    technique = get_technique(f"https://w3id.org/PaN/ESRFET/{version}/#XRF")
+    assert technique in all_techniques
+
+
 def test_ontology_version():
     version = get_ontology_version()
     pattern = r"^v\d+\.\d+\.\d+$"