+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed ontology base IRI and all entity IRIs from http://purl.org/pan-science/ESRFET to https://w3id.org/PaN/ESRFET.
 - Save technique metadata in HDF5 using the `/NXobject/identifierNAME-field` fields.
 - Versioned IRI stored for the NeXus-compliant format instead of IRI.
+- `get_technique_metadata` and `get_techniques` now support both technique names and IRIs as input identifiers.
 
 ### Added